organize arm components
